+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- (Experimental) `Formal\ORM\Matching::property()` to do cross aggregate matching
+
 ## 5.0.0 - 2025-06-04
 
 ### Changed